added fix for bug in accelleration code
previously, 0 velocity was obtained by having 0 velocity in the motor and either backward or forward direction. this caused bugs when moving from a sop position. now to be in 0 velocity, the motor has to be in forward direction

TODO: right now the code works for small accelerations. its behaviour for big accelerations is cumbersome and should be improved.